+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Animatie en object toegevoegd aan radio gebruik
 - Banden reparatie mogelijkheid toegevoegd
 - Jerrycan toegevoegd
+- whitelist signed-in spelers kunnen geen job groep meer aanmaken/joinen
 
 ### Changed
 - Mogelijkheid om verschillende verkoop aantallen per item te bepalen bij cornersell